pyproject.toml
[project]
name = "pdf-flattening"
dynamic = ["version"]
description = "Convert PDFs to image-only PDFs (one raster image per page)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pypdfium2",
    "Pillow",
    "tqdm",
]

[project.scripts]
pdf-flatten = "pdf_flattening.cli:main"

[tool.setuptools.dynamic]
version = { file = "VERSION" }

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"